Fixed more type incoherence.

+++ internal/app/mart_app_server.go
@@ -297,11 +297,11 @@ type orderResponse struct {
 
 type withdrawalsResponse struct {
 	Order       string    `json:"order"`
-	Sum         int64     `json:"sum"`
+	Sum         float64   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
 
 type balanceWithdrawRequest struct {
-	Order string `json:"order"`
-	Sum   int64  `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float64 `json:"sum"`
 }
